tempCommit

+++ src/main/java/ch/heigvd/sprint0/controller/AdminController.java
@@ -6,6 +6,7 @@ import ch.heigvd.sprint0.model.Category;
 import ch.heigvd.sprint0.service.IArticleCategoryService;
 import ch.heigvd.sprint0.service.IArticleService;
 import ch.heigvd.sprint0.service.ICategoryService;
+import ch.heigvd.sprint0.service.SessionService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Controller;
@@ -17,8 +18,11 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.multipart.MultipartFile;
 
 import javax.imageio.ImageIO;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 import java.io.InputStream;
+import java.net.http.HttpRequest;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
@@ -34,18 +38,22 @@ public class AdminController {
 
     private final ICategoryService categoryService;
 
+    private final SessionService sessionService;
+
     @Value("${server.tomcat.upload-dir}")
     private String uploadPath;
 
     @Autowired
-    public AdminController(IArticleService articleService, IArticleCategoryService articleCategoryService, ICategoryService categoryService) {
+    public AdminController(IArticleService articleService, IArticleCategoryService articleCategoryService, ICategoryService categoryService, SessionService sessionService) {
         this.articleService = articleService;
         this.articleCategoryService = articleCategoryService;
         this.categoryService = categoryService;
+        this.sessionService = sessionService;
     }
 
     @GetMapping("/admin")
-    public String admin(Model model) {
+    public String admin(Model model, HttpServletRequest request, HttpServletResponse response) {
+        checkAdminAccess(request, response);
         List<Article> articles = articleService.findAll();
 
         model.addAttribute("articles", articles);
@@ -56,6 +64,7 @@ public class AdminController {
     public String adminArticle(Model model, @RequestParam(name = "id", required = false) String id,
                                @RequestParam(name = "error", required = false) String error,
                                @RequestParam(name = "error_msg", required = false) String errorMsg) {
+        //checkAdminAccess(request, response);
         if(id != null && !id.chars().allMatch(Character::isDigit))
             return admin(model);
 
@@ -87,8 +96,8 @@ public class AdminController {
     @PostMapping("/admin/article")
     public String articleSubmit(@ModelAttribute Article article, Model model,
                                 @RequestParam(value = "image", required = false) MultipartFile image,
-                                @RequestParam(value = "article_category_list", required = false) String article_category_list) {
-
+                                @RequestParam(value = "article_category_list", required = false) String article_category_list, HttpServletRequest request, HttpServletResponse response) {
+        checkAdminAccess(request, response);
         Optional<Article> articleWithSameDescription = articleService.findByDescription(article.getDescription());
         // C'est une modification d'article
         if(article.getId() != null) {
@@ -167,7 +176,8 @@ public class AdminController {
     }
 
     @GetMapping("/admin/article/delete")
-    public String articleDelete(@RequestParam(name = "id") String id) {
+    public String articleDelete(@RequestParam(name = "id") String id, HttpServletRequest request, HttpServletResponse response) {
+        checkAdminAccess(request, response);
         // L'id existe ?
         Optional<Article> article = articleService.findById(Integer.parseInt(id));
         article.ifPresent(value -> articleService.deleteById(value.getId()));
@@ -177,7 +187,8 @@ public class AdminController {
 
     @GetMapping("/admin/categories")
     public String adminCategories(Model model,
-                                  @RequestParam(name = "error", required = false) String error) {
+                                  @RequestParam(name = "error", required = false) String error, HttpServletRequest request, HttpServletResponse response) {
+        checkAdminAccess(request, response);
         List<Category> categories = categoryService.findAll();
         if(!categories.isEmpty()) {
             model.addAttribute("categories", categories);
@@ -189,7 +200,8 @@ public class AdminController {
     }
 
     @PostMapping("/admin/categories")
-    public String categorySubmit(Model model, @ModelAttribute Category category) {
+    public String categorySubmit(Model model, @ModelAttribute Category category, HttpServletRequest request, HttpServletResponse response) {
+        checkAdminAccess(request, response);
         // Vérifier que le nom de la catégorie existe pas déjà
         Optional<Category> cat = categoryService.findByName(category.getNameCategory());
         if(cat.isPresent()) {
@@ -201,7 +213,8 @@ public class AdminController {
     }
 
     @GetMapping("/admin/categories/confirmDeletion")
-    public String confirmCategoryDeletion(Model model, @RequestParam(name = "id") String id) {
+    public String confirmCategoryDeletion(Model model, @RequestParam(name = "id") String id, HttpServletRequest request, HttpServletResponse response) {
+        checkAdminAccess(request, response);
         // L'id existe ?
         Optional<Category> category = categoryService.findByName(id);
         if(category.isPresent()) {
@@ -225,7 +238,8 @@ public class AdminController {
     }
 
     @GetMapping("/admin/categories/delete")
-    public String categoryDelete(Model model, @RequestParam(name = "id") String id) {
+    public String categoryDelete(Model model, @RequestParam(name = "id") String id, HttpServletRequest request, HttpServletResponse response) {
+        checkAdminAccess(request, response);
         // L'id existe ?
         Optional<Category> category = categoryService.findByName(id);
         if(category.isPresent()) {
@@ -233,5 +247,18 @@ public class AdminController {
         }
 
         return adminCategories(model,null);
+    }
+
+
+    private void checkAdminAccess(HttpServletRequest request, HttpServletResponse response) {
+        String[] userData = sessionService.checkLogin(request);
+        if(!(userData != null && userData[1].equals("admin"))) {    // Etre logué + rôle admin pour accéder à la page
+            try {
+                response.sendRedirect("./");                     // Redirection, access denied
+            } catch (IOException e) {
+                e.printStackTrace();
+            }
+        }
+
     }
 }
